fix: extract legacy ipc
