+++ crates/server/Cargo.toml
@@ -6,5 +6,8 @@ edition = "2018"
 license = "AGPL-3.0"
 publish = false
 
+[features]
+legacy = ["rendezvous-common/legacy"]
+
 [dependencies]
 rendezvous-common = { path = "../common" }
